% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_if_unique}
\alias{add_if_unique}
\title{Add row to dataframe, or remove it if it already exists}
\usage{
add_if_unique(original, append, x, y)
}
\arguments{
\item{original}{The input dataframe}

\item{append}{A dataframe containing the same columns as `original` and 1 row}
}
\value{
If the row in `append` exists in `original` it will be removed, otherwise it will be added to the dataframe that is returned.
}
\description{
Add row to dataframe, or remove it if it already exists
}
