% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_picker.R
\name{add_peaks}
\alias{add_peaks}
\title{Add peak indicators at the given indices}
\usage{
add_peaks(df, indices)
}
\arguments{
\item{df}{Dataframe with dataset (e.g. x- and y-values of a spectrum)}

\item{indices}{Vector of integer indices indicating peak positions}
}
\value{
tbl_df with an additional column indicating whether the row is a peak
}
\description{
Add peak indicators at the given indices
}
\examples{
library(tidyverse)
tibble(x1 = seq(0.1, 9, 0.01), y1 = sin(x1)) \%>\%
  add_peaks(c(148,776))
}
