% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_gradient}
\alias{find_gradient}
\title{Local gradient at given index}
\usage{
find_gradient(x, y, ind)
}
\arguments{
\item{x}{Vector of x-values}

\item{y}{Vector of y-values}

\item{ind}{Index to calculate the gradient at}
}
\value{
numeric
}
\description{
Local gradient at given index
}
