% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_peak}
\alias{find_peak}
\title{Returns the index of the nearest peak}
\usage{
find_peak(x, y, ind)
}
\arguments{
\item{x}{Vector of x-values}

\item{y}{Vector of y-values}

\item{ind}{Index to start peak search}
}
\value{
integer index of peak
}
\description{
Calculates the gradient at the given index and then travels along the line in the upwards direction until a peak is found
}
