% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generic_df}
\alias{generic_df}
\title{Return a standard dataframe with two columns (x and y)}
\usage{
generic_df(df, x, y)
}
\arguments{
\item{df}{Dataframe to be converted}

\item{x}{Column containing x-coordinates}

\item{y}{Column containing y-coordinates}
}
\value{
A standard tibble with the given data in two columns (x and y)
}
\description{
Return a standard dataframe with two columns (x and y)
}
