% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{peak_integrate}
\alias{peak_integrate}
\title{area_picker}
\usage{
peak_integrate(df, x, y)
}
\arguments{
\item{df}{}
}
\description{
area_picker
}
\examples{
df <- echem_read(system.file('extdata/cv/cv_example.txt', package = 'osc'))
df <- area_picker(df)
}
