% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakr.R
\docType{package}
\name{peakr}
\alias{peakr}
\alias{peakr-package}
\title{\code{peakr} package}
\description{
Interactive tools for analysis of small datasets
}
\details{
See the README on
\href{https://cran.r-project.org/package=peakr/README.html}{CRAN}
or \href{https://github.com/emiltb/peakr#readme}{GitHub}
}
