% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_picker.R
\name{plot_peaks}
\alias{plot_peaks}
\title{Plot a dataset contain defined peaks}
\usage{
plot_peaks(df, x, y)
}
\arguments{
\item{df}{Any dataframe with e.g. a spectrum or similar. It must contain a column 'peak' containing TRUE/FALSE values indicating the presence of peaks.}

\item{x}{The values plotted on the x-axis (e.g. wavelength)}

\item{y}{The values plotted on the y-axis (e.g. intensity/counts)}
}
\value{
A ggplot
}
\description{
Plot a dataset contain defined peaks
}
\examples{
library(tidyverse)
tibble(x1 = seq(0.1, 9, 0.01), y1 = sin(x1)) \%>\%
  add_peaks(c(148,776)) \%>\%
  plot_peaks(x1, y1)
}
