% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{plot_smooth}
\alias{plot_smooth}
\title{Plot results of data smooting}
\usage{
plot_smooth(df, x, y)
}
\arguments{
\item{df}{A tibble containing the data to pick peaks in}

\item{x}{Column containing the x-values}

\item{y}{Column containing the y-values}
}
\value{
A plot displaying the orignal and smoothed data, as well as the residual
}
\description{
Plot results of data smooting
}
