% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{add_pick}
\alias{add_pick}
\title{Add peak indicators at the given indices}
\usage{
add_pick(data, indices)
}
\arguments{
\item{data}{A tibble containing the data to pick peaks in}

\item{indices}{Indices of the data to mark as peaks}
}
\value{
tbl_df with an additional column indicating whether the row is a peak
}
\description{
Add peak indicators at the given indices
}
\examples{
library(tibble)
library(dplyr)
tibble(x1 = seq(0.1, 9, 0.01), y1 = sin(x1)) \%>\%
  add_pick(c(148,776))
}
