% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{plot_pick}
\alias{plot_pick}
\title{Plot a dataset contain defined peaks}
\usage{
plot_pick(data, x, y)
}
\arguments{
\item{x}{The values plotted on the x-axis (e.g. wavelength)}

\item{y}{The values plotted on the y-axis (e.g. intensity/counts)}

\item{df}{Any dataframe with e.g. a spectrum or similar. It must contain a column 'peak' containing TRUE/FALSE values indicating the presence of peaks.}
}
\value{
A ggplot
}
\description{
Plot a dataset contain defined peaks
}
\examples{
library(tibble)
library(dplyr)

tibble(x1 = seq(0.1, 9, 0.01), y1 = sin(x1)) \%>\%
  add_pick(c(148,776)) \%>\%
  plot_pick(x1, y1)
}
